requirements.txt
#requirements.txt
snowflake-connector-python
streamlit
pandas
requests
matplotlib

